requirements.txt
requests>=1.1
python-dateutil
pycountry


